Use streaming endpoint in list_microgrid_components_data
The streaming endpoint supports requests for historical data. This
replaces the list endpoint with the streaming endpoint. The names in the
client are for now unchanged to not break user code. They might be
revised when live streaming will also be supported by the service.

The list endpoint will most likely be deprecated and removed at some point.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>