Add support for states in client and cli tool
States can now be requested via the client and are provided through the
flat iterator. They can be identified via their category `state`,
`warning` and `error`, respectively. Each individual state is provided
as its own sample.

Support for states is also added to the CLI tool via the `--states`
flag. As of now requests for states without any metrics are not yet
supported by the service.

Signed-off-by: cwasicki <126617870+cwasicki@users.noreply.github.com>